% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastierinstall.R
\docType{package}
\name{beastierinstall}
\alias{beastierinstall}
\title{\code{beastierinstall}: A package to install BEAST2.}
\description{
\code{beastierinstall} allows to install and uninstall BEAST2,
a popular Bayesian phylogenetics tool, using
an R interface.
}
\details{
\code{beastierinstall} in a non-CRAN extension of the
\code{beastier} package. With \code{beastier} one can work with BEAST2
using an R interface.
}
