% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_beautier_cache.R
\name{clear_beautier_cache}
\alias{clear_beautier_cache}
\title{Clear the \link{beautier} cache}
\usage{
clear_beautier_cache(beautier_folder = beautier::get_beautier_folder())
}
\arguments{
\item{beautier_folder}{the path to
the \link{beautier} temporary files folder}
}
\value{
Nothing.
}
\description{
Clear the \link{beautier} cache.
}
\author{
Richèl J.C. Bilderbeek
}
